Get rid of codegenlinear.h
